[Feat] 이벤트를 진행하기 위한 조건 체크
